+++ CHANGELOG.md
@@ -4,7 +4,10 @@ miru.ink / Changelog
 ## Version 0.0.24 - Experimental
 - Add: New `vitest` + `@vue/test-utils` testing framework.
 - Add: New `badge` variant on `AlertMessage` component.
+- Add: New `AlertDialog` component, similar to the `ConfirmDialog` one + basic story demonstration.
+- Add: New `DividerStd` component with different variants + basic story demonstration.
 - Update: Use `info` color on `AlertMessage` component + Fix dark stylings.
+- Update: Minor styling changes on `DialogStd` component.
 - Update: package.json dependencies.
 
 ## Version 0.0.23 - Experimental